Criação das exeções das classes de négocios

+++ src/main/java/br/edu/ufape/poo/listadecompras/negocios/NegocioConta.java
@@ -6,17 +6,31 @@ import java.util.Optional;
 import br.edu.ufape.poo.listadecompras.dados.IRepositorioContas;
 import br.edu.ufape.poo.listadecompras.negocios.cadastro.InterfaceCadastroConta;
 import br.edu.ufape.poo.listadecompras.negocios.entidade.Conta;
+import br.edu.ufape.poo.listadecompras.negocios.excecoes.ContaNaoEncontradaException;
+import br.edu.ufape.poo.listadecompras.negocios.excecoes.NaoEncontradoPeloEmailException;
+import br.edu.ufape.poo.listadecompras.negocios.excecoes.NaoEncontradoPeloIdException;
 
 public class NegocioConta implements InterfaceCadastroConta{
 
     //Fazer as exceções
     private IRepositorioContas repositorioContas;
+    
 
-    public List<Conta> procurarContaEmail(String email){
+    public List<Conta> procurarContaEmail(String email) 
+    throws NaoEncontradoPeloEmailException{
+        
+        if(repositorioContas.findByEmail(email) == null || repositorioContas.findByEmail(email).isEmpty()){
+            throw  new NaoEncontradoPeloEmailException(email);
+        }
+       
         return repositorioContas.findByEmail(email);
     }
 
-    public void salvarConta(Conta entity){
+    public void salvarConta(Conta entity)
+    throws ContaNaoEncontradaException{
+        if(repositorioContas.findById(entity.getId()) == null){
+            throw new ContaNaoEncontradaException(entity);
+        }
         repositorioContas.save(entity);
     }
 
@@ -24,15 +38,27 @@ public class NegocioConta implements InterfaceCadastroConta{
         return repositorioContas.findAll();
     }
 
-	public void removerConta(Long id){
+	public void removerConta(Long id)
+    throws NaoEncontradoPeloIdException{
+        if(repositorioContas.findById(id) == null){
+            throw new NaoEncontradoPeloIdException(id);
+        }
         repositorioContas.deleteById(id);
     }
 
-	public void removerConta(Conta entity){
+	public void removerConta(Conta entity)
+    throws ContaNaoEncontradaException{
+        if(repositorioContas.findById(entity.getId()) == null){
+            throw new ContaNaoEncontradaException(entity);
+        }
         repositorioContas.delete(entity);
     }
 
-	public Optional<Conta> localizarContaId(long id){
+	public Optional<Conta> localizarContaId(long id)
+    throws NaoEncontradoPeloIdException{
+        if(repositorioContas.findById(id) == null){
+            throw new NaoEncontradoPeloIdException(id);
+        }
         return repositorioContas.findById(id);
     }
 }
